Rename flexcontainer package -> flexbox.
