+++ redwood-flex-container/src/commonMain/kotlin/app/cash/redwood/flexbox/Measurable.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.flexcontainer
+package app.cash.redwood.flexbox
 
 public open class Measurable {
   /**
