+++ redwood-flex-container/src/commonMain/kotlin/app/cash/redwood/flexbox/MeasureResult.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.flexcontainer
+package app.cash.redwood.flexbox
 
 /**
  * Holds the state produced by calling [FlexContainer.measure].
