+++ redwood-flex-container/src/commonMain/kotlin/app/cash/redwood/flexbox/properties.kt
@@ -15,7 +15,7 @@
  */
 @file:Suppress("MemberVisibilityCanBePrivate")
 
-package app.cash.redwood.flexcontainer
+package app.cash.redwood.flexbox
 
 import kotlin.jvm.JvmInline
