+++ redwood-layout-composeui/src/main/kotlin/app/cash/redwood/layout/composeui/ComposeColumn.kt
@@ -17,7 +17,7 @@ package app.cash.redwood.layout.composeui
 
 import androidx.compose.runtime.Composable
 import app.cash.redwood.LayoutModifier
-import app.cash.redwood.flexcontainer.FlexDirection
+import app.cash.redwood.flexbox.FlexDirection
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.layout.api.MainAxisAlignment
