+++ redwood-layout-uiview/build.gradle
@@ -1,4 +1,4 @@
-import app.cash.redwood.buildsupport.FlexContainerHelpers
+import app.cash.redwood.buildsupport.FlexboxHelpers
 
 apply plugin: 'org.jetbrains.kotlin.multiplatform'
 apply plugin: 'com.vanniktech.maven.publish'
@@ -11,7 +11,7 @@ kotlin {
 
   sourceSets {
     commonMain {
-      kotlin.srcDir(FlexContainerHelpers.get(tasks, 'app.cash.redwood.layout.uiview'))
+      kotlin.srcDir(FlexboxHelpers.get(tasks, 'app.cash.redwood.layout.uiview'))
       dependencies {
         api projects.redwoodLayoutSchema.widget
         implementation projects.redwoodFlexContainer
