+++ redwood-layout-uiview/src/commonMain/kotlin/app/cash/redwood/layout/uiview/UIViewColumn.kt
@@ -16,7 +16,7 @@
 package app.cash.redwood.layout.uiview
 
 import app.cash.redwood.LayoutModifier
-import app.cash.redwood.flexcontainer.FlexDirection
+import app.cash.redwood.flexbox.FlexDirection
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.layout.api.MainAxisAlignment
