+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/view/ViewColumn.kt
@@ -18,7 +18,7 @@ package app.cash.redwood.layout.view
 import android.content.Context
 import android.view.View
 import app.cash.redwood.LayoutModifier
-import app.cash.redwood.flexcontainer.FlexDirection
+import app.cash.redwood.flexbox.FlexDirection
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.layout.api.MainAxisAlignment
