+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/view/ViewFlexContainer.kt
@@ -22,14 +22,14 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.HorizontalScrollView
 import androidx.core.widget.NestedScrollView
-import app.cash.redwood.flexcontainer.AlignItems
-import app.cash.redwood.flexcontainer.FlexContainer
-import app.cash.redwood.flexcontainer.FlexDirection
-import app.cash.redwood.flexcontainer.JustifyContent
-import app.cash.redwood.flexcontainer.MeasureResult
-import app.cash.redwood.flexcontainer.MeasureSpec as RedwoodMeasureSpec
-import app.cash.redwood.flexcontainer.Size
-import app.cash.redwood.flexcontainer.isHorizontal
+import app.cash.redwood.flexbox.AlignItems
+import app.cash.redwood.flexbox.FlexContainer
+import app.cash.redwood.flexbox.FlexDirection
+import app.cash.redwood.flexbox.JustifyContent
+import app.cash.redwood.flexbox.MeasureResult
+import app.cash.redwood.flexbox.MeasureSpec as RedwoodMeasureSpec
+import app.cash.redwood.flexbox.Size
+import app.cash.redwood.flexbox.isHorizontal
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.Overflow
 import app.cash.redwood.layout.api.Padding
